mb/google/brya: Update PCH power cycle related durations
The voltage rail discharge times have been measured, so therefore
the boot time on a cold boot when the CSE must go through a global reset
and thus a trip to S5 can be optimized. Select the lowest applicable
value for each PchPmSlp UPD that can be used with these measurements.

This is programmed in the baseboard because the measured discharge times
leave (what should be) plenty of margin for variants to also not violate
any power sequencing guidelines from the PDG.

BUG=b:184799383
TEST=verified time in S5 during a global reset is ~1s instead of 4s

Change-Id: Ia373c47b3967d68ddac21707c6eb4565d9d6519e
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/57892
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: EricR Lai <ericr_lai@compal.corp-partner.google.com>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
GitOrigin-RevId: dee834aafcccf79f9d6c3319ead6212fced004a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3195458
Tested-by: Copybara Service <copybara-worker-blackhole@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>