google/trogdor: Always initialize eDP bridge I2C QUP firmware
In CB:52662 when MIPI display support was added, we accidentally changed
the code flow for eDP displays such that i2c_init() will no longer be
called when display_init_required() is false. This is a problem because
on this platform, i2c_init() does not just prepare the I2C controller
for firmware use, it also loads firmware to the controller that makes it
behave like an I2C device in the first place -- a step that the kernel
cannot later do on its own if the firmware didn't already do it.
Skipping this initialization means the I2C controller becomes unusable
to the kernel.

This patch fixes the issue by making the i2c_init() unconditional again.

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: Ie4546c31d87d91113eeef7dc7a18599a87e6d6eb
Reviewed-on: https://review.coreboot.org/c/coreboot/+/58026
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: f67f2b0d4ca4d1d72fa077d69abef916c8c3a38c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3195470
Tested-by: Copybara Service <copybara-worker-blackhole@google.com>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>