mb/google/dedede/var/bugzzy: Update GPP_D5 configuration
As we checked the panel doesn't display firmware screen if we hold
GPP_D5(TOUCHSCREEN_RESET) low on bugzzy. It's because of that bugzzy
uses the built-in touch screen on the panel, the panel seems like
under reset state by the TOUCHSCREEN_RESET signal.
This change sets default GPP_D5 level to high for bugzzy.

BUG=b:None
BRANCH=dedede
TEST=built and verified bugzzy showed firmware screen

Signed-off-by: Seunghwan Kim <sh_.kim@samsung.corp-partner.google.com>
Change-Id: Ia4bc584b3101637ed1d49c10d253b2069e5c8199
Reviewed-on: https://review.coreboot.org/c/coreboot/+/58073
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Edward Doan <edoan@google.com>
Reviewed-by: Karthik Ramasubramanian <kramasub@google.com>
GitOrigin-RevId: bd4487c86911a62138352c310c2e4bbdbd054453
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3215341
Tested-by: Rob Barnes <robbarnes@google.com>
Tested-by: Copybara Service <copybara-worker-blackhole@google.com>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-by: Edward Doan <edoan@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>