lib/hardwaremain: change type of "complete" element in boot_state struct
A signed bitfield with a length of 1 bit can only have the values 0 and
-1. Assigning a 1 ends up behaving as expected, but it's not the
semantically correct thing to do there. Changing the type of the element
to an unsigned bitfield with a length of 1 would fix that, but since
this is used as a boolean value, just change it to bool type.

Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Change-Id: Ib796be5ecf9cce99d89aa96256ae838a1fb424b7
Reviewed-on: https://review.coreboot.org/c/coreboot/+/58076
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
GitOrigin-RevId: 56da0b79ad6f3a96f5c14405d8622b7ed10aa1f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3215300
Tested-by: Rob Barnes <robbarnes@google.com>
Tested-by: Copybara Service <copybara-worker-blackhole@google.com>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Commit-Queue: Rob Barnes <robbarnes@google.com>