coreboot tables: Add type-c port info to coreboot table
This change adds type-c port information for USB Type-C ports to the
coreboot table.  This allows depthcharge to know the usb2 and usb3
port number assignments for each available port, as well as the SBU
and data line orientation for the board.

BUG=b:149830546
TEST='emerge-volteer coreboot chromeos-bootimage' and verify it builds
successfully.  Cherry-pick CL to enable this feature for volteer,
flash and boot volteer2 to kernel, log in and check cbmem for type-c
info exported to the payload:
  localhost ~ # cbmem -c | grep type-c
  added type-c port0 info to cbmem: usb2:9 usb3:1 sbu:0 data:0
  added type-c port1 info to cbmem: usb2:4 usb3:2 sbu:1 data:0

Signed-off-by: Nick Vaccaro <nvaccaro@google.com>
Change-Id: I60b01ce44999ef5c4a72d08ac819bfe3b8581943
Reviewed-on: https://review.coreboot.org/c/coreboot/+/57069
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
GitOrigin-RevId: a62b41819ceda4108b3a97cf72f55fc470347d58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3215301
Tested-by: Rob Barnes <robbarnes@google.com>
Tested-by: Copybara Service <copybara-worker-blackhole@google.com>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Commit-Queue: Rob Barnes <robbarnes@google.com>