soc/amd/common/block/i2c: implement proper read_resource
Before this patch the reservation of the MMIO region of the I2C
controllers was done in the LPC controller PCI device despite the I2C
controllers already being devices in the devicetree. This patch
implements this functionality as read_resources function of the I2C
device instead. This will only reserve the memory when the I2C devices
are enabled in devicetree which is a change from the previous behavior.

Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Change-Id: I77a587aefcc061ce284b50d6b18b7b79915e82d2
Reviewed-on: https://review.coreboot.org/c/coreboot/+/58307
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Karthik Ramasubramanian <kramasub@google.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: c2cee06b4efb8992e625ca4554df17c7ef457930
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3227784
Tested-by: Dossym Nurmukhanov <dossym@chromium.org>
Reviewed-by: Dossym Nurmukhanov <dossym@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>