cpu/x86: Introduce and use `CPU_X86_LAPIC`
With using a Kconfig option to add the x86 LAPIC support code to the
build, there's no need for adding the corresponding directory to subdirs
in the CPU/SoC Makefile. Comparing which CPU/SoC Makefiles added
(cpu/)x86/mtrr and (cpu/)x86/lapic before this and the corresponding
MTRR code selection patch and having verified that all platforms
added the MTRR code on that patch shows that soc/example/min86 and
soc/intel/quark are the only platforms that don't end up selecting the
LAPIC code. So for now the default value of CPU_X86_LAPIC is chosen as y
which gets overridden to n in the Kconfig of the two SoCs mentioned
above.

Change-Id: I6f683ea7ba92c91117017ebc6ad063ec54902b0a
Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/44228
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-by: Marshall Dawson <marshalldawson3rd@gmail.com>
GitOrigin-RevId: 02164027b2c9678bd4f5616e0ec34694da351a2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3246489
Tested-by: Copybara Service <copybara-worker-blackhole@google.com>
Tested-by: Rob Barnes <robbarnes@google.com>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Reviewed-by: Bhanu Prakash Maiya <bhanumaiya@google.com>
Reviewed-by: Andrew McRae <amcrae@google.com>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Andrew McRae <amcrae@google.com>